Restrict delete commnet by owner

+++ comment/views.py
@@ -206,13 +206,22 @@ class CommentDetail(APIView):
             if Question.objects.filter(id=self.kwargs['question_id']):
                 if Comment.objects.filter(question=self.kwargs['question_id']):
                     comment = self.get_object(pk)
-                    comment.delete()
+                    comment_owner = comment.created_by
+                    if comment_owner == request.user:
+                        comment.delete()
+                        return Response(
+                            {
+                                "status": status.HTTP_204_NO_CONTENT,
+                                "message": "Comment successfully deleted."
+                            },
+                            status=status.HTTP_204_NO_CONTENT
+                        )
                     return Response(
                         {
-                            "status": status.HTTP_204_NO_CONTENT,
-                            "message": "Comment successfully deleted."
+                            "status": status.HTTP_403_FORBIDDEN,
+                            "error": 'You cannot delete this comment.'
                         },
-                        status=status.HTTP_204_NO_CONTENT
+                        status=status.HTTP_403_FORBIDDEN
                     )
             return Response(
                 {
